payments_manager_flask/requirements.txt
Flask==1.1.2
python-dotenv==0.19.0

# DB
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.20
Flask-Migrate==2.5.3
Flask-Script==2.0.6
mysqlclient==2.0.1

# CORS
Flask-Cors==3.0.9

# AMQP
pika==1.1.0

# validation
marshmallow==3.11.1

# date-utils
maya==0.6.1

# json-serializer
simplejson==3.17.2

# PEP8
black==20.8b1

# Tests
pytest==6.2.3
Flask-Testing==0.6.2
Werkzeug==0.16.1